refactor(ingestion): move progress registry and tracking mixin into metadata/ingestion/progress package
