+++ ingestion/src/metadata/ingestion/source/database/query_parser_source.py
@@ -30,7 +30,7 @@ from metadata.ingestion.ometa.ometa_api import OpenMetadata
 from metadata.ingestion.source.connections import test_connection_common
 from metadata.utils.helpers import get_start_and_end, retry_with_docker_host
 from metadata.utils.logger import ingestion_logger
-from metadata.utils.progress_tracking import ProgressTrackingMixin
+from metadata.ingestion.progress.tracking import ProgressTrackingMixin
 from metadata.utils.ssl_manager import get_ssl_connection
 
 logger = ingestion_logger()
